feat(commands): lazy load commands
This should help minimize startup time for `ng` commands.
