+++ packages/angular-cli/commands/lint.ts
@@ -1,11 +1,11 @@
 const Command = require('../ember-cli/lib/models/command');
-import LintTask from '../tasks/lint';
 
 export default Command.extend({
   name: 'lint',
   description: 'Lints code in existing project',
   works: 'insideProject',
   run: function () {
+    const LintTask = require('../tasks/lint').default;
     const lintTask = new LintTask({
       ui: this.ui,
       project: this.project
