+++ packages/angular-cli/commands/serve.ts
@@ -1,14 +1,8 @@
-import * as denodeify from 'denodeify';
-const assign = require('lodash/assign');
-const SilentError = require('silent-error');
 const PortFinder = require('portfinder');
 const Command = require('../ember-cli/lib/models/command');
-import ServeWebpackTask from '../tasks/serve-webpack';
-import {Version} from '../upgrade/version';
 
 PortFinder.basePort = 49152;
 
-const getPort = <any>denodeify(PortFinder.getPort);
 const defaultPort = process.env.PORT || 4200;
 
 export interface ServeTaskOptions {
@@ -113,70 +107,7 @@ const ServeCommand = Command.extend({
   ],
 
   run: function(commandOptions: ServeTaskOptions) {
-    if (commandOptions.environment === '') {
-      if (commandOptions.target === 'development') {
-        commandOptions.environment = 'dev';
-      }
-      if (commandOptions.target === 'production') {
-        commandOptions.environment = 'prod';
-      }
-    }
-
-    // Check angular version.
-    Version.assertAngularVersionIs2_3_1OrHigher(this.project.root);
-    commandOptions.liveReloadHost = commandOptions.liveReloadHost || commandOptions.host;
-
-    return this._checkExpressPort(commandOptions)
-      .then(this._autoFindLiveReloadPort.bind(this))
-      .then((opts: ServeTaskOptions) => {
-        commandOptions = assign({}, opts, {
-          baseURL: this.project.config(commandOptions.target).baseURL || '/'
-        });
-
-        const serve = new ServeWebpackTask({
-          ui: this.ui,
-          project: this.project,
-        });
-
-        return serve.run(commandOptions);
-      });
-  },
-
-  _checkExpressPort: function(commandOptions: ServeTaskOptions) {
-    return getPort({ port: commandOptions.port, host: commandOptions.host })
-      .then((foundPort: number) => {
-
-        if (commandOptions.port !== foundPort && commandOptions.port !== 0) {
-          throw new SilentError(`Port ${commandOptions.port} is already in use.`);
-        }
-
-        // otherwise, our found port is good
-        commandOptions.port = foundPort;
-        return commandOptions;
-
-      });
-  },
-
-  _autoFindLiveReloadPort: function(commandOptions: ServeTaskOptions) {
-    return getPort({ port: commandOptions.liveReloadPort, host: commandOptions.liveReloadHost })
-      .then((foundPort: number) => {
-
-        // if live reload port matches express port, try one higher
-        if (foundPort === commandOptions.port) {
-          commandOptions.liveReloadPort = foundPort + 1;
-          return this._autoFindLiveReloadPort(commandOptions);
-        }
-
-        // port was already open
-        if (foundPort === commandOptions.liveReloadPort) {
-          return commandOptions;
-        }
-
-        // use found port as live reload port
-        commandOptions.liveReloadPort = foundPort;
-        return commandOptions;
-
-      });
+    return require('./serve.run').default.call(this, commandOptions);
   }
 });
